+++ trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/media/PlatformMedia.kt
@@ -1,0 +1,7 @@
+package net.folivo.trixnity.client.media
+
+import net.folivo.trixnity.utils.ByteArrayFlow
+
+interface PlatformMedia : ByteArrayFlow {
+    fun transformByteArrayFlow(transformer: (ByteArrayFlow) -> ByteArrayFlow): PlatformMedia
+}