+++ trixnity-client/trixnity-client-repository-indexeddb/src/jsMain/kotlin/net/folivo/trixnity/client/store/repository/indexeddb/IndexedDBOlmAccountRepository.kt
@@ -18,10 +18,7 @@ internal class IndexedDBOlmAccountRepository(
     companion object {
         const val objectStoreName = "olm_account"
         fun VersionChangeTransaction.migrate(database: Database, oldVersion: Int) {
-            when {
-                oldVersion < 1 ->
-                    createIndexedDBMinimalStoreRepository(database, objectStoreName)
-            }
+            if (oldVersion < 1) createIndexedDBMinimalStoreRepository(database, objectStoreName)
         }
     }
 }